refactor(ui): optimize icon imports and clean up social link mappings
- Move `components/icons.tsx` to `components/ui/icons.tsx` and update all corresponding import paths
- Replace static `BRAND_ICONS.<icon>` usages with direct named component imports where applicable
