+++ components/layout/footer.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link";
 import Image from "next/image";
 import { Mail } from "lucide-react";
-import { BRAND_ICONS } from "../icons";
+import { BRAND_ICONS } from "@/components/ui/icons";
 
 const footerLinks = {
   community: {
