+++ components/ui/dual-cta.tsx
@@ -2,7 +2,7 @@ import Link from "next/link";
 import { Calendar, Users, ExternalLink } from "lucide-react";
 import { Button } from "@/components/ui/button";
 import { getEventLinks, getSocialLinks, getContributeLinks } from "@/constants/social-links";
-import { BRAND_ICONS } from "@/components/icons";
+import { BRAND_ICONS } from "@/components/ui/icons";
 
 export function DualCTA() {
   const eventLinks = getEventLinks();
@@ -24,7 +24,7 @@ export function DualCTA() {
           </p>
           <div className="flex flex-wrap gap-2">
             {[...eventLinks, ...socialLinks].map((link) => {
-              const Icon = BRAND_ICONS[link.name.toLowerCase()];
+              const Icon = BRAND_ICONS[link.id];
 
               return (
                 <Button key={link.id} variant={link.type === "primary" ? "primary" : "outline"} size="sm" asChild>
@@ -54,7 +54,7 @@ export function DualCTA() {
               <Link href="/get-involved">Get involved</Link>
             </Button>
             {contributeLinks.map((link) => {
-              const Icon = BRAND_ICONS[link.name.toLowerCase()];
+              const Icon = BRAND_ICONS[link.id];
               return (
                 <Button key={link.id} variant="outline" size="sm" asChild>
                   <a href={link.url} target="_blank" rel="noopener noreferrer">
