+++ components/ui/social-feed.tsx
@@ -3,7 +3,7 @@
 import Image from "next/image";
 import { ExternalLink } from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { BRAND_ICONS } from "@/components/icons";
+import { BRAND_ICONS, InstagramIcon } from "@/components/ui/icons";
 import { getAllSocialLinks } from "@/constants/social-links";
 import {
   getCoverImage,
@@ -68,7 +68,7 @@ function InstagramPostCard({ post }: { post: InstagramPostWithId }) {
       <div className="p-6 flex flex-col flex-1">
         <div className="flex flex-wrap items-center gap-x-2 gap-y-1 mb-4">
           <span className="text-[#E4405F]">
-            <BRAND_ICONS.instagram className="h-5 w-5" />
+            <InstagramIcon className="h-5 w-5" />
           </span>
           <span className="text-xs text-muted-foreground uppercase tracking-wide font-medium shrink-0">Instagram</span>
           <span className="text-xs text-muted-foreground/50 shrink-0">·</span>
@@ -87,7 +87,7 @@ function InstagramPostCard({ post }: { post: InstagramPostWithId }) {
           rel="noopener noreferrer"
           className="mt-auto inline-flex items-center text-sm text-amber-dark font-medium hover:text-amber-dark/80 transition-colors"
         >
-          <BRAND_ICONS.instagram className="mr-2 h-4 w-4" />
+          <InstagramIcon className="mr-2 h-4 w-4" />
           View on Instagram
           <ExternalLink className="ml-1 h-3 w-3" />
         </a>
